requirements.txt
mysql-connector-python
pyfcm
python-socketio
requests
